move scale/shift calc to math utils
